+++ siso-ios/Coordinator/IntegratedPage.swift
@@ -1,6 +1,7 @@
 import Foundation
 import model // UserProfileServer를 사용하기 위해 필요
 import chat
+import network
 
 public enum IntegrationPage: Hashable, Identifiable {
     // Auth
