+++ siso-ios/Matching/Coordinator/MatchingCoordinatorDelegate.swift
@@ -7,6 +7,8 @@
 
 import SwiftUI
 import model
+import network
+
 
 public protocol MatchingCoordinatorDelegate: AnyObject {
