+++ siso-ios/Network/LoginNetwork.swift
@@ -50,7 +50,7 @@ public final actor LoginNetworkManager: Sendable {
             switch response.result {
             case .success(var token):
                 // 2. KeyChainManager를 사용해 RefreshToken 토큰 저장
-                self?.keychain.save(token: token.refreshToken, for: "accessToken")
+                self?.keychain.save(token: token.accessToken, for: "accessToken")
                 self?.keychain.save(token: token.refreshToken, for: "refreshToken")
                 if !token.hasProfile  {
                     token.registrationStatus = "REGISTER" // 동의 항목 이동
@@ -97,7 +97,7 @@ public final actor LoginNetworkManager: Sendable {
             .value
         
         // 새로 받은 토큰을 다시 저장
-        keychain.save(token: response.token.refreshToken, for: "accessToken")
+        keychain.save(token: response.token.accessToken, for: "accessToken")
         keychain.save(token: response.token.refreshToken, for: "refreshToken")
         print("refreshToken : \(response.token.refreshToken)")
         if !response.token.hasProfile {
